pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cli-patterns"
version = "0.1.0"
description = "A framework for building consistent and beautiful command-line interfaces"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "CLI Patterns Team", email = "team@cli-patterns.dev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Shells",
    "Topic :: Terminals",
]
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0",
    "pytest>=7.0",
    "pytest-asyncio",
    "pre-commit",
    "black",
    "ruff",
]

[project.urls]
Homepage = "https://github.com/cli-patterns/cli-patterns"
Documentation = "https://cli-patterns.readthedocs.io"
Repository = "https://github.com/cli-patterns/cli-patterns"
Issues = "https://github.com/cli-patterns/cli-patterns/issues"

[project.scripts]
cli-patterns = "cli_patterns.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "UP007", # Use X | Y for type annotations (requires Python 3.10+)
    "UP045", # Use X | None for type annotations (requires Python 3.10+)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for component interactions",
    "parser: Tests for parser system components",
    "executor: Tests for execution/subprocess components",
    "design: Tests for design system components",
    "slow: Tests that take significant time to run",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.1",
]
